Update fixtures folder
