+++ packages/dom/test/unit/lib/fixtures/toBeInDocTestComponent.tsx
@@ -1,0 +1,9 @@
+import { ReactElement } from "react";
+
+export function TestComponent(): ReactElement {
+    return (
+      <div>
+        <button>click me</button>
+      </div>
+    );
+  }
