+++ packages/dom/test/unit/lib/fixtures/toContainElementTestComponent.tsx
@@ -1,0 +1,20 @@
+import { ReactElement } from "react";
+
+export function ContainsTestComponent(): ReactElement {
+  return (
+    <span data-testid="grandparent">
+      <span data-testid="parent">
+        <span data-testid="child">
+          <span>
+              <span>
+                  <span>
+                      <span data-testid="deep-child"></span>
+                  </span>
+              </span>
+          </span>
+        </span>
+      </span>
+      <svg data-testid="svg-element"></svg>
+    </span>
+  );
+}
